Pull extractor code out into modules
Currently we have lots of logic for implementing the Extrator role. To
make this clear pull the logic out into separate modules and explicitly
mark it as WIP.
