+++ src/v0/mod.rs
@@ -7,6 +7,7 @@
 //! [BIP-174]: <https://github.com/bitcoin/bips/blob/master/bip-0174.mediawiki>
 
 mod error;
+mod extractor;
 mod map;
 #[cfg(feature = "miniscript")]
 pub mod miniscript;
@@ -21,11 +22,11 @@ use bitcoin::key::{PrivateKey, PublicKey};
 use bitcoin::secp256k1::{Message, Secp256k1, Signing};
 use bitcoin::sighash::{EcdsaSighashType, SighashCache};
 use bitcoin::transaction::{Transaction, TxOut};
-use bitcoin::{ecdsa, Amount, FeeRate};
+use bitcoin::{ecdsa, Amount};
 
 use crate::error::write_err;
 use crate::prelude::*;
-use crate::v0::error::{ExtractTxError, IndexOutOfBoundsError, SignError};
+use crate::v0::error::{IndexOutOfBoundsError, SignError};
 use crate::v0::map::{Global, Input, Map, Output};
 use crate::Error;
 
@@ -154,97 +155,6 @@ impl Psbt {
             outputs: vec![Default::default(); output_len],
         };
         Ok(psbt)
-    }
-
-    /// The default `max_fee_rate` value used for extracting transactions with [`extract_tx`]
-    ///
-    /// As of 2023, even the biggest overpayers during the highest fee markets only paid around
-    /// 1000 sats/vByte. 25k sats/vByte is obviously a mistake at this point.
-    ///
-    /// [`extract_tx`]: Psbt::extract_tx
-    pub const DEFAULT_MAX_FEE_RATE: FeeRate = FeeRate::from_sat_per_vb_unchecked(25_000);
-
-    /// An alias for [`extract_tx_fee_rate_limit`].
-    ///
-    /// [`extract_tx_fee_rate_limit`]: Psbt::extract_tx_fee_rate_limit
-    pub fn extract_tx(self) -> Result<Transaction, ExtractTxError> {
-        self.internal_extract_tx_with_fee_rate_limit(Self::DEFAULT_MAX_FEE_RATE)
-    }
-
-    /// Extracts the [`Transaction`] from a [`Psbt`] by filling in the available signature information.
-    ///
-    /// ## Errors
-    ///
-    /// `ExtractTxError` variants will contain either the [`Psbt`] itself or the [`Transaction`]
-    /// that was extracted. These can be extracted from the Errors in order to recover.
-    /// See the error documentation for info on the variants. In general, it covers large fees.
-    pub fn extract_tx_fee_rate_limit(self) -> Result<Transaction, ExtractTxError> {
-        self.internal_extract_tx_with_fee_rate_limit(Self::DEFAULT_MAX_FEE_RATE)
-    }
-
-    /// Extracts the [`Transaction`] from a [`Psbt`] by filling in the available signature information.
-    ///
-    /// ## Errors
-    ///
-    /// See [`extract_tx`].
-    ///
-    /// [`extract_tx`]: Psbt::extract_tx
-    pub fn extract_tx_with_fee_rate_limit(
-        self,
-        max_fee_rate: FeeRate,
-    ) -> Result<Transaction, ExtractTxError> {
-        self.internal_extract_tx_with_fee_rate_limit(max_fee_rate)
-    }
-
-    /// Perform [`extract_tx_fee_rate_limit`] without the fee rate check.
-    ///
-    /// This can result in a transaction with absurdly high fees. Use with caution.
-    ///
-    /// [`extract_tx_fee_rate_limit`]: Psbt::extract_tx_fee_rate_limit
-    pub fn extract_tx_unchecked_fee_rate(self) -> Transaction { self.internal_extract_tx() }
-
-    #[inline]
-    fn internal_extract_tx(self) -> Transaction {
-        let mut tx: Transaction = self.global.unsigned_tx;
-
-        for (vin, psbtin) in tx.input.iter_mut().zip(self.inputs.into_iter()) {
-            vin.script_sig = psbtin.final_script_sig.unwrap_or_default();
-            vin.witness = psbtin.final_script_witness.unwrap_or_default();
-        }
-
-        tx
-    }
-
-    #[inline]
-    fn internal_extract_tx_with_fee_rate_limit(
-        self,
-        max_fee_rate: FeeRate,
-    ) -> Result<Transaction, ExtractTxError> {
-        let fee = match self.fee() {
-            Ok(fee) => fee,
-            Err(Error::MissingUtxo) =>
-                return Err(ExtractTxError::MissingInputValue { tx: self.internal_extract_tx() }),
-            Err(Error::NegativeFee) => return Err(ExtractTxError::SendingTooMuch { psbt: self }),
-            Err(Error::FeeOverflow) =>
-                return Err(ExtractTxError::AbsurdFeeRate {
-                    fee_rate: FeeRate::MAX,
-                    tx: self.internal_extract_tx(),
-                }),
-            _ => unreachable!(),
-        };
-
-        // Note: Move prevents usage of &self from now on.
-        let tx = self.internal_extract_tx();
-
-        // Now that the extracted Transaction is made, decide how to return it.
-        let fee_rate =
-            FeeRate::from_sat_per_kwu(fee.to_sat().saturating_mul(1000) / tx.weight().to_wu());
-        // Prefer to return an AbsurdFeeRate error when both trigger.
-        if fee_rate > max_fee_rate {
-            return Err(ExtractTxError::AbsurdFeeRate { fee_rate, tx });
-        }
-
-        Ok(tx)
     }
 
     /// Combines this [`Psbt`] with `other` PSBT as described by BIP 174.
@@ -740,10 +650,11 @@ mod tests {
     use bitcoin::secp256k1::{All, SecretKey};
     use bitcoin::transaction::{self, OutPoint, Sequence, Transaction, TxIn, TxOut};
     use bitcoin::Network::Bitcoin;
-    use bitcoin::{ScriptBuf, Witness};
+    use bitcoin::{FeeRate, ScriptBuf, Witness};
 
     use super::*;
     use crate::serialize::{Deserialize, Serialize};
+    use crate::v0::error::ExtractTxError;
     use crate::{io, raw};
 
     #[track_caller]
